pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arcadiaforge"
version = "0.1.0"
description = "Autonomous coding agent harness demo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "claude-code-sdk>=0.0.25",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "prompt_toolkit>=3.0.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
]

[project.scripts]
autonomous-agent = "arcadiaforge.cli.autonomous_agent:main"
checkpoint-cli = "arcadiaforge.cli.checkpoint_cli:main"
events-cli = "arcadiaforge.cli.events_cli:main"
feature-cli = "arcadiaforge.cli.feature_cli:main"
metrics-cli = "arcadiaforge.cli.metrics_cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"arcadiaforge.prompts" = ["*.md", "*.txt"]
